Split pooling functionality into its own class so that it doesn't interfere with application servers which do their own pooling and transaction management.
